AutomationFramework_V1/requirements.txt
pyspark>=3.0.0
pandas>=1.0.0
numpy>=1.18.0
PyYAML>=5.1
pytest>=6.0.0
matplotlib>=3.0.0
tabulate>=0.8.7
colorama>=0.4.3
jira>=2.0.0
python-dateutil>=2.8.0
requests>=2.25.0
tqdm>=4.50.0
dash>=2.5.0
dash-bootstrap-components>=1.2.0
plotly>=5.8.0
sqlite3-api>=0.1.0 